+++ controllers/product.controller.js
@@ -1,4 +1,8 @@
+const { request } = require('http');
 const Product = require('../models/product.model');
+const Workshops = require('../models/workshops.model');
+const Category = require('../models/category.model');
+const { response } = require('express');
 
 // POST /product: register a new product
 exports.postRegisterProduct = async (request, response) => {
@@ -34,6 +38,7 @@ exports.getProducts = async (request, response) => {
     }
 }
 
+// GET: Consult one product
 exports.getOneProduct = async (request, response) => {
     try {
         const id = request.params.idProduct;
@@ -42,6 +47,34 @@ exports.getOneProduct = async (request, response) => {
         response.status(200).json(product);
     } catch (err) {
         // Error
-        response.status(500).json({ message: "Failed to consult products", err });
+        response.status(500).json({ message: "Failed to consult product", err });
+    }
+}
+
+// PUT: modify an excisting product
+exports.updateProduct = async (request, response) => {
+    try {        
+        const id = request.params.idProduct;
+        
+        const idTaller = await Workshops.getId(request.body.nombreTaller); 
+        const idCategoria = await Category.getId(request.body.Categoria);
+        
+        const productData = {
+            idTaller: idTaller,
+            Nombre: request.body.Nombre,
+            PrecioUnitario: request.body.PrecioUnitario,
+            idCategoria: idCategoria,
+            Descripcion: request.body.Descripcion,
+            imagen: request.body.imagen,
+            Disponible: request.body.Disponible,
+        };
+
+        await Product.update(id, productData); 
+
+        // Success
+        response.status(200).json({ message: "Product modified successfully" }); 
+    } catch (err) {
+        // Error
+        response.status(500).json({ message: "Failed to modify a product", err });
     }
 }